# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sicherheitsdienst-Tool API
  version: 1.0.0
  description: API v1 für Auth, Mitarbeiter, Sites, Schichten, Zuweisungen, Zeiterfassung, Vorfälle, Benachrichtigungen.
servers:
  - url: http://localhost:3000/api/v1
  - url: http://localhost:3001/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ShiftStatus:
      type: string
      enum: [PLANNED, ACTIVE, COMPLETED, CANCELLED]
    Shift:
      type: object
      properties:
        id:
          type: string
          format: uuid
        siteId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        requiredEmployees:
          type: integer
          format: int32
        requiredQualifications:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/ShiftStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, location, startTime, endTime, requiredEmployees, status, createdAt, updatedAt]
...
  responses:
    BadRequest:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: BAD_REQUEST
            message: Anfrageparameter sind ungültig.
    Unauthorized:
      description: Nicht autorisiert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: UNAUTHORIZED
            message: Token fehlt oder ist ungültig.
    Forbidden:
      description: Verboten
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: FORBIDDEN
            message: Keine Berechtigung.
    NotFound:
      description: Nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: NOT_FOUND
            message: Ressource nicht gefunden.
...
    TooManyRequests:
      description: Zu viele Anfragen
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: TOO_MANY_REQUESTS
            message: Rate-Limit erreicht.
...
paths:
...
  /shifts/{id}:
    get:
      operationId: getShift
      summary: Schicht abrufen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Schicht
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
              example:
                id: "b1b2b3b4-1111-2222-3333-444455556666"
                siteId: "7c1f2e94-5678-4a12-9cde-abcdef123456"
                title: "Nachtschicht"
                description: "Wachdienst Haupteingang"
                location: "Messedamm 22"
                startTime: "2024-09-01T22:00:00Z"
                endTime: "2024-09-02T06:00:00Z"
                requiredEmployees: 2
                requiredQualifications: ["Sicherheitskraft"]
                status: "PLANNED"
                createdAt: "2024-08-25T10:00:00Z"
                updatedAt: "2024-08-25T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      operationId: updateShift
      summary: Schicht aktualisieren
      description: 'RBAC: ADMIN oder DISPATCHER. Feature-Flag: Bei EMAIL_NOTIFY_SHIFTS=true werden E-Mails an zugewiesene Mitarbeiter versendet.'
      x-feature-flag: EMAIL_NOTIFY_SHIFTS
      x-required-roles:
        - ADMIN
        - DISPATCHER
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftUpdate'
            example:
              title: "Nachtschicht (verlängert)"
              endTime: "2024-09-02T07:00:00Z"
      responses:
        '200':
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
              example:
                id: "b1b2b3b4-1111-2222-3333-444455556666"
                siteId: "7c1f2e94-5678-4a12-9cde-abcdef123456"
                title: "Nachtschicht (verlängert)"
                description: "Wachdienst Haupteingang"
                location: "Messedamm 22"
                startTime: "2024-09-01T22:00:00Z"
                endTime: "2024-09-02T07:00:00Z"
                requiredEmployees: 2
                requiredQualifications: ["Sicherheitskraft"]
                status: "PLANNED"
                createdAt: "2024-08-25T10:00:00Z"
                updatedAt: "2024-09-01T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      operationId: deleteShift
      summary: Schicht löschen
      description: 'RBAC: ADMIN. Feature-Flag: Bei EMAIL_NOTIFY_SHIFTS=true werden E-Mails an zugewiesene Mitarbeiter versendet.'
      x-feature-flag: EMAIL_NOTIFY_SHIFTS
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Schicht gelöscht (Antwort enthält Basisdaten)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      title:
                        type: string
              example:
                success: true
                message: 'Schicht erfolgreich gelöscht'
                data:
                  id: 'b1b2b3b4-1111-2222-3333-444455556666'
                  title: 'Nachtschicht'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /shifts/{id}/assignments:
    post:
      operationId: assignUserToShift
      summary: Mitarbeiter zuweisen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
            example:
              userId: "a3f1c5d2-1234-4a56-9b8c-0d1e2f3a4b5c"
              status: "ASSIGNED"
      responses:
        '201':
          description: Zuweisung erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
              example:
                id: "c0ffee00-aaaa-bbbb-cccc-111122223333"
                userId: "a3f1c5d2-1234-4a56-9b8c-0d1e2f3a4b5c"
                shiftId: "b1b2b3b4-1111-2222-3333-444455556666"
                status: "ASSIGNED"
                assignedAt: "2024-08-25T10:00:00Z"
                updatedAt: "2024-08-25T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      operationId: removeUserFromShift
      summary: Mitarbeiter-Zuweisung entfernen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Entfernt
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /shifts/{id}/clock-in:
    post:
      operationId: clockIn
      summary: Arbeitsbeginn buchen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockInRequest'
            example:
              at: "2024-09-01T22:05:00Z"
              location: "Messedamm 22"
              notes: "Start an Tor 3"
      responses:
        '200':
          description: Zeiteintrag erstellt (Warnungen möglich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTimeEntry'
              example:
                success: true
                message: "Clock-in erfasst"
                data:
                  id: "d34db33f-aaaa-bbbb-cccc-111122223333"
                  userId: "a3f1c5d2-1234-4a56-9b8c-0d1e2f3a4b5c"
                  startTime: "2024-09-01T22:05:00Z"
                  endTime: null
                  breakTime: null
                  startLocation: "Messedamm 22"
                  endLocation: null
                  notes: "Start an Tor 3"
                  createdAt: "2024-09-01T22:05:00Z"
                  updatedAt: "2024-09-01T22:05:00Z"
                warnings: ["WARN_REST_PERIOD_LT_11H"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /shifts/{id}/clock-out:
    post:
      operationId: clockOut
      summary: Arbeitsende buchen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockOutRequest'
            example:
              at: "2024-09-02T06:05:00Z"
              breakTime: 15
              location: "Messedamm 22"
              notes: "Ende Tor 3"
      responses:
        '200':
          description: Zeiteintrag aktualisiert (Warnungen möglich)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTimeEntry'
              example:
                success: true
                message: "Clock-out erfasst"
                data:
                  id: "d34db33f-aaaa-bbbb-cccc-111122223333"
                  userId: "a3f1c5d2-1234-4a56-9b8c-0d1e2f3a4b5c"
                  startTime: "2024-09-01T22:05:00Z"
                  endTime: "2024-09-02T06:05:00Z"
                  breakTime: 15
                  startLocation: "Messedamm 22"
                  endLocation: "Messedamm 22"
                  notes: "Ende Tor 3"
                  createdAt: "2024-09-01T22:05:00Z"
                  updatedAt: "2024-09-02T06:05:00Z"
                warnings: ["WARN_SHIFT_GT_10H"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...